refactor: move to `qr_code` under `src`
