+++ src/qr_code/admin.py
