+++ src/qr_code/apps.py
@@ -1,6 +1,7 @@
 from django.apps import AppConfig
 
 
-class SrcConfig(AppConfig):
+class QrCodeConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'src'
+    name = 'src.qr_code'
+    label = 'qr_code'
