+++ src/qr_code/models.py
