+++ src/qr_code/static/css/style.css
