+++ src/qr_code/templates/base.html
