+++ src/qr_code/views.py
